Make selection menu for Oauth provider
