+++ app/helpers/rp/main_helper.rb
@@ -1,0 +1,2 @@
+module Rp::MainHelper
+end
