+++ app/views/rp/main/index.html.erb
@@ -1,0 +1,11 @@
+<h1>ユーザ登録／ログイン</h1>
+<p>ご指定の操作にはユーザ登録／ログインが必要です</p>
+
+<p>
+お持ちのアカウントでユーザ登録／ログインをお願いします
+</p>
+<ul>
+  <li><a href=<%= @gp.to_s %>?<%= @p1.to_s %>&<%= @p2.to_s %>>Googleで登録／ログイン</a></li>
+  <li><a href="#">Face Bookで登録／ログイン</a></li>
+  <li><a href="#">Twitterで登録／ログイン</a></li>
+</ul>
