+++ test/controllers/rp/main_controller_test.rb
@@ -1,0 +1,9 @@
+require 'test_helper'
+
+class Rp::MainControllerTest < ActionDispatch::IntegrationTest
+  test "should get index" do
+    get rp_main_index_url
+    assert_response :success
+  end
+
+end
